+++ lib/feature/data/dto/payment_detail.dto.dart
@@ -1,6 +1,6 @@
 class PaymentDetail {
   String? description;
-  double? value;
+  int? value;
 
   PaymentDetail({this.description, this.value});
